Address review feedback
- Alphabetize driversLicenseItemState in AddEditItemState and CipherItemState
- Move driversLicenseFixture to BitwardenSdk+VaultFixtures
- Make date display helper a private static func (drop file-scope globals)
- Convert DriversLicenseItemStateTests to Swift Testing
