+++ BitwardenShared/UI/Vault/PreviewContent/BitwardenSdk+VaultFixtures.swift
@@ -346,6 +346,56 @@ extension CipherView {
         )
     }
 
+    /// A `CipherView` fixture populated with a fully-filled driver's license, used to verify the
+    /// state round-trip preserves all 11 fields including the raw date strings.
+    static func driversLicenseFixture() -> CipherView {
+        CipherView(
+            id: "1",
+            organizationId: nil,
+            folderId: nil,
+            collectionIds: [],
+            key: nil,
+            name: "Bitwarden",
+            notes: nil,
+            type: .driversLicense,
+            login: nil,
+            identity: nil,
+            card: nil,
+            secureNote: nil,
+            sshKey: nil,
+            bankAccount: nil,
+            driversLicense: DriversLicenseView(
+                firstName: "Bit",
+                middleName: "W",
+                lastName: "Warden",
+                dateOfBirth: "1989-08-01",
+                licenseNumber: "D1234567",
+                issuingCountry: "United States",
+                issuingState: "California",
+                issueDate: "2019-08-01",
+                expirationDate: "2029-08-01",
+                issuingAuthority: "DMV",
+                licenseClass: "C",
+            ),
+            passport: nil,
+            favorite: false,
+            reprompt: .none,
+            organizationUseTotp: false,
+            edit: true,
+            permissions: nil,
+            viewPassword: true,
+            localData: nil,
+            attachments: nil,
+            attachmentDecryptionFailures: nil,
+            fields: nil,
+            passwordHistory: nil,
+            creationDate: Date(year: 2023, month: 11, day: 5),
+            deletedDate: nil,
+            revisionDate: Date(year: 2023, month: 11, day: 5),
+            archivedDate: nil,
+        )
+    }
+
     static func loginFixture(
         archivedDate: Date? = nil,
         attachments: [AttachmentView]? = nil,
